requirements.txt
langchain
langgraph
langchain-openai
langchain-community
streamlit
arxiv
duckduckgo-search
python-dotenv
